+++ extensions/git-id-switcher/src/test/runTests.ts
@@ -29,6 +29,7 @@ import { runPathSeparatorTests } from './pathSeparator.test';
 import { runDisplayLimitsTests } from './displayLimits.test';
 import { runSshAgentParsingTests } from './sshAgentParsing.test';
 import { runSyncCheckerTests } from './syncChecker.test';
+import { runErrorTests } from './errors.test';
 
 async function main(): Promise<void> {
   console.log('╔════════════════════════════════════════════╗');
@@ -110,6 +111,9 @@ async function main(): Promise<void> {
 
     // Run sync checker tests (profile vs git config comparison)
     await runSyncCheckerTests();
+
+    // Run error classes and validation types tests
+    await runErrorTests();
 
     console.log('╔════════════════════════════════════════════╗');
     console.log('║   🎉 All Security Tests Passed!            ║');
